requirements.txt
Flask==2.3.3
pandas==2.0.3
openpyxl==3.1.2
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
sentry-sdk[flask]==1.38.0
pytest==7.4.3